Change from int to float

+++ f-101/chapter5-numbers-and-math/5.3_challenge_exponent.py
@@ -1,7 +1,7 @@
 # Write a script called exponent.py that receives two numbers from the user and displays the first number raised to the power of the second number.
 
 base = float(input("Enter a base: "))
-exponent = int(input("Enter an exponent: "))
+exponent = float(input("Enter an exponent: "))
 exponentiation = base**exponent
 
 result = f"{base} to the power of {exponent} = {exponentiation}"
